chore: add support to version diff from semver
It adds a utility function to check if the version satisfies the range
rather than use the `satisfies` direct from the `semver` package.
This function try coerce the version to a valid semver before compare.
From the tests, versions such as v1 and 1 pass. So, from now versions
following this patterns works as semver.