chore: verifies the module system before import config file
It imports the package.json from projectDirectory to check the module system,
ECMAScript Modules or CommonJS. If the  module system is ESM then just import
the eventcatalog.config.js. Otherwise, make a copy of eventcatalog.config.js
to the tmpdir/eventcatalog.config.mjs and import it.
It removes the convertEsmToCjs used before.