fix/ownerReference (#1078)
* fix: change ownerReference to union of string and object

It replaces the `reference` from the ownerReference to a union of
string and an object with id and collection. The string refers to
the owner id. The object with id and collection keep compatibility
with `reference` in case someone use like that.

It adds a collection utility to get the owner from the onwerReference.

Fix #1034

* Create wild-items-shout.md

---------

Co-authored-by: David Boyne <boyneyy123@gmail.com>