fix(core): remove empty index page to prevent blank docs page (#1170)
* fix(core): remove empty index page to prevent blank docs page

This removes the empty markdown file that were previously added by
`ensureAstroCollectionNotEmpty` when no index page was specified by
the user. This introduce `default-docs.mdx` as fallback to prevent
blank /docs page. This approach ensures a better UX without redirecting
to the first collection item, as an initial empty project may not have
any items to navigate to.

Closes #1130.

* Create eleven-parrots-grab.md

---------

Co-authored-by: David Boyne <boyneyy123@gmail.com>