feat: add protocol check to tile component's href handling
It add a function to check if the href starts with a protocol (e.g., http, https, ftp).
If the href passed is a valid absolute URLs, it is used directly in the anchor element.
If not, the `buildUrl` utility is used to create the URL from the passed `href`.

Fix #1201